pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "ffmpeg>=1.4",
    "imageio-ffmpeg>=0.6.0",
    "langchain-core>=1.2.12",
    "langchain-openai>=1.1.9",
    "numpy>=2.4.1",
    "onnxruntime>=1.19.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "scipy>=1.7.0",
    "sherpa-onnx>=1.12.23",
    "sounddevice>=0.5.5",
    "torch>=2.0.0",
    "torchaudio>=2.10.0",
    "uvicorn>=0.40.0",
    "whisper>=1.1.10",
]
